Rename params to param
