+++ zkevm-circuits/src/state_circuit.rs
@@ -2,7 +2,8 @@
 
 // TODO: clean up module structure.
 pub(super) mod constraint_builder;
-pub(super) mod params; // TODO: rename to param
+pub(super) mod fixed_table;
+pub(super) mod param;
 
 pub(crate) mod state;
 pub use state::StateCircuit;
