+++ zkevm-circuits/src/state_circuit/state.rs
@@ -1,5 +1,5 @@
 use super::constraint_builder::ConstraintBuilder;
-use super::params::N_LIMBS_ACCOUNT_ADDRESS;
+use super::param::N_LIMBS_ACCOUNT_ADDRESS;
 use crate::{
     evm_circuit::{
         param::N_BYTES_WORD,
